+++ src/MainScene.ts
@@ -1,47 +1,18 @@
-import { Botist, IAdapter } from './Botist';
-import {
-    ITextMessage,
-    IImageMessage,
-    IMessage,
-} from './Message.t';
+import { IMessage } from './Message.t';
 import { Response } from './Response';
+import { MessageMiddleware, IMessageMiddleware, ISubscribers } from './Middlewares/Message';
 import { IEvent } from './Events/Event';
 
-interface IMessageHandlers {
-    text: ITextMessageHandler[];
-    image: IImageMessageHandler[];
-}
-
-export type MessageCallback<M> = (msg: M, res: Response, next: () => void) => void;
-
-interface ITextMessageHandler {
-    text: string | RegExp;
-    callback: MessageCallback<ITextMessage>;
-}
-
-interface IImageMessageHandler {
-    callback: MessageCallback<IImageMessage>;
-}
-
-export interface IScene {
-    subscribe(): void;
+export interface IScene extends IMessageMiddleware {
+    subscribers: ISubscribers;
     enter(msg: IMessage, res: Response, event: IEvent): void;
     leave(msg: IMessage, res: Response, event: IEvent): void;
-    text(text: string | RegExp, callback: MessageCallback<ITextMessage>): void;
-    onMessage(adapter: IAdapter, msg: IMessage, startHandlerIndex?: number): void;
 }
 
-export abstract class MainScene implements IScene {
-    private messageHandlers: IMessageHandlers = {
-        text: [],
-        image: [],
-    };
-
-    constructor(private botist: Botist) {
-        this.subscribe();
+export abstract class MainScene extends MessageMiddleware implements IScene {
+    public get subscribers(): ISubscribers {
+        return this._subscribers;
     }
-
-    public abstract subscribe(): void;
 
     /**
      * Called when scene is activating.
@@ -52,32 +23,4 @@ export abstract class MainScene implements IScene {
      * Called when scene is deactivating.
      */
     public leave(_msg: IMessage, _res: Response, _event: IEvent) {}
-
-    public text(text: string | RegExp, callback: MessageCallback<ITextMessage>): void {
-        this.messageHandlers.text.push({
-            text,
-            callback,
-        });
-    }
-
-    public onMessage(adapter: IAdapter, msg: IMessage, startHandlerIndex: number = 0): void {
-        for (let i = startHandlerIndex; i < this.messageHandlers[msg.type].length; i++) {
-            const handler = this.messageHandlers[msg.type][i];
-            if (msg.type === 'text') {
-                const handlerText = (handler as ITextMessageHandler).text;
-                if (
-                    typeof handlerText === 'string' && msg.text !== handlerText ||
-                    handlerText instanceof RegExp && !msg.text.match(handlerText)
-                ) {
-                    continue;
-                }
-            }
-
-            const res = new Response(this.botist, adapter, msg);
-            handler.callback.call(null, msg, res, () => {
-                this.onMessage(adapter, msg, i + 1);
-            });
-            break;
-        }
-    }
 }