+++ src/index.ts
@@ -9,11 +9,13 @@ export {
     ITextMessage,
     IImageMessage,
 } from './Message.t';
-export { MainScene } from './MainScene';
+export { MainScene, IScene } from './MainScene';
 export { Scene } from './Scene';
 export { Scenario } from './Scenario';
 export { Messenger } from './Messenger/Adapter';
 export { Telegram } from './Telegram/Adapter';
 export { Response as IResponse } from './Response';
+export { MessageMiddleware, ISubscriber, ISubscriberCallback } from './Middlewares/Message';
+export { SubscriberContext } from './SubscriberContext';
 export { IEvent } from './Events/Event';
 export * from './Events/SceneEvents';